refactor: 폴더 구조 정리 mainLayout폴더 만들어서 분리했고, router파일 분리
