--- src/component/MainContents.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Main = () => {
-  return <div></div>;
-};
-
-export default Main;
